+++ old/chec.py
@@ -1,0 +1,2 @@
+import menu_tree_builder
+print(f"[DEBUG] Chargé depuis : {menu_tree_builder.__file__}")
