+++ old/tettt.py
